libnet/driverapi: Introduce `EndpointDriver`
This new interface embeds the Driver interface and defines all four
methods related to Endpoint management: CreateEndpoint, Join, Leave and
EndpointOperInfo.

Drivers implementing this new interface also return a new capability
named 'EndpointDriver'. When libnetwork wants to call one of the
`EndpointDriver` method, it first checks whether the network driver has
this capability. If it doesn't, it early returns.

Because of that, libnet's `Endpoint` isn't necessarily backed by a
driver Endpoint. In such case, it represents an endpoint-less 'network
attachment'.

Also note that the logic in `(*Sandbox).removeEndpointRaw()` is modified
to look for an endpoint by its ID instead of comparing struct instances
are the same. The old logic is quite weird as libnetwork tends to create
multiple copies of the same 'object' into memory. Why / how this worked
is unclear, but this commit made a few tests fail exactly because of
that.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>