api: Make EnableIPv6 optional
Currently, starting dockerd with
`--default-network-opt=bridge=com.docker.network.enable_ipv6=true` has
no effect as `NetworkCreateRequest.EnableIPv6` is a basic bool.

This change uses a Rust-like `Option` type to mark the field as
optional. If clients don't specify it, the default-network-opt will be
applied if it's specified, otherwise it defaults to false.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>