daemon: move LB creation out of libnet
Prior to this commit, libnet `Controller.NewNetwork` was in charge of
creating a load-balancer sandbox and connecting it to the network being
created, if `NetworkOptionLBEndpoint` was passed.

This was effectively hiding details about 'load-balancers', which are
ultimately networking concepts, from the daemon. But this came at a
rather high cost: this makes libnet self-driven rather than being solely
daemon-driven. (that's not the only case.)

This might seems innocuous, but it's in fact preventing future
refactorings like merging `CreateEndpoint` and `Join` methods, or
revamping the concurrency model.

Since libnet's logic to decide whether a network can be deleted is
influenced by the presence of a load-balancer, it still knows what a
'load-balancer' is. Only the create responsibility is handed off to the
daemon.

The property `loadBalancerIP` in libnet's Network is removed. There's no
need to keep marshaling it for preserving the ability to downgrade since
only `overlay` networks could get this property set, and `overlay`
networks are never persisted in the first place.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>