libnet: default_gwbridge: drop chan procGwNetwork
This channel was used in `defaultGwNetwork` like a global mutex to
ensure no two concurrent calls to `setupDefaultGW` would race against
each other to create the 'docker_gwbridge'. This was added in [1] to fix
the issue reported in [2].

Since commit 78479b1, calls made to NewNetwork with the same name are
properly serialized by libnet's Controller. If concurrent calls are
made, the first call winning the race will create the network and other
calls will return a 'NetworkNameError'.

`defaultGwNetwork` is updated to detect that error and retrieve the
Network from the datastore.

[1]: https://github.com/moby/libnetwork/pull/899
[2]: https://github.com/moby/moby/issues/19677

Signed-off-by: Albin Kerouanton <albinker@gmail.com>