daemon: move docker_gwbridge mgmt out of libnet
Prior to this commit, `(*Endpoint).sbJoin()` was connecting the sandbox
to the `docker_gwbridge` network when needed. When doing so, the
`docker_gwbridge` network was created if it didn't exist yet.

This whole nested flow was weird, and made libnetwork in charge of
deciding when Sandbox, Network and Endpoint operations would happen.
This is last case where libnetwork is driving itself.

This commit transfers the responsibility of connecting a sandbox to the
'docker_gwbridge' nw to the daemon.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>