daemon/cluster: remove Cluster.getRequestContext()
This method was added in 534a90a99367af6f6bba1ddcc7eb07506e41f774 as
part of adding the Swarm cluster backend, and later updated in commit
85b1fdf15ce2ad1b373748554d3aa218e2eb5a5f to use a swarmRequestTimeout
const for the timeout.

Nothing in this utility depends on the Cluster struct, and the abstraction
makes it appear as more than it is, which is just a wrapper for
context.WithTimeout().

Let's remove the abstraction to make it less magical.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>