WIP: Try to recover from log parsing errors
When reading logs (from json and local drivers), the data may be
corrupted in some cases.

This can happen, as an example, due to power failures and data not being
flushed to disk.
ext4 paired with IoT type devices which are often not battery backed are
especially prone to this due to ext4's defaulting to use deferred
allocation which causes only metadata to be journaled.
In this case we'll end up with a bunch of null bytes written to the log
file.

Another case can just be older log files created with buggy versions of
docker.

I don't think we should treat these as critical errors that should
abort log reading, especially since it would require manual cleanup to
ever be able to read the logs (or recreating the container entirely).
Instead this change will advance the reader to the next position after
the failed offset.
This may cause a bunch of churn trying to parse by advancing one byte at
a time. Another option may be to just skip to the next file entirely
instead of trying to recover a bad file, but I think that's a bit more
drastic and we could end up losing a lot more data (at the API layer)
than is desirable.

Currently the only way to detect that dockerd is having a problem
parsing a log file is to wire up a tracer as we don't have a means of
passing warning messages to the client.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>