pkg/reexec: don't mix syscall and golang.org/x/sys package
commit 069fdc8a083cb1663e4f86fe3fd9b9a1aebc3e54 changed most uses of
the syscall package to switch utsname from unsigned to signed (see
069fdc8a083cb1663e4f86fe3fd9b9a1aebc3e54). Those don't seem to be
impacting the code used here, so either stdlib or golang.org/x/sys/unix
should work for this case.

I chose stdlib's syscall package for this case, in case we'd decide to
move this package to a separate module (and want to limit its dependencies).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>