libnet: EndpointInfo: replace DisableGatewayService
`(*Sandbox).needDefaultGw` was going into great length to meticulously
exclude all networks except non-internal overlays, and remote netdrivers
(when the remote didn't set DisableGatewayService).

`DisableGatewayService` is now replaced with a new Endpoint property:
`RequireDefaultGateway`, and all the cruft in `needDefaultGw` is
simplified. The new property is set during the Join operation, and
restored networks get it set automtically when loaded from the datastore.

The old endpoint's property `disableGatewayService` is marshaled into
the on-disk state to make sure downgrades are still possible.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>