Add option com.docker.network.bridge.host_binding_ip
Existing option com.docker.network.bridge.host_binding_ipv4 accepts
IPv6 addresses - so, added a synonym without the 'v4'.

It's an error to specify both.

The existing handling is surprising - apart from accepting v6 in
an option named v4, "0.0.0.0" means "any v4 or v6 address" but "::"
means "any v6 address". So, added an integration test to illustrate
that and make sure it's not "fixed" in passing.

Signed-off-by: Rob Murray <rob.murray@docker.com>