libnet/ds: rename Address into Path
When Classic Swarm was still supported, libnet datastore could use a
variety of remote backends. The `Address` field was meant to store a URI
to one of these remote backend.

Now that only boltdb backend is supported, there's no need to keep
naming that field an 'address'. Rename it to `Path`.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>