libnet/ds: merge boltdb.Config into Config
The outer-struct `Config` was providing no value, and everything from
the inner and the outer-struct was needed to created a boltdb storage
backend. Keeping both structs wasn't providing any value.

This could have been made the other way around: merge the outer into the
inner, ie. Config into boltdb.Config. But `NewConfig` is currently part
of the public API, and it's the only way for libnetwork consumers (ie.
outside  of moby) to build the datastore config. So instead of keeping
`NewConfig` in the datastore package, and move all the config into the
internal `boltdb.Config`, it seems better to move everything into the
public struct `datastore.Config`.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>