pkg/dmesg: deprecate, and use internal utility instead
This package was originally added in 46833ee1c353c247e3ef817a08d5a35a2a43bdf3
for use in the devicemapper graphdriver. The devicemapper graphdriver was
deprecated and has been removed. The only remaining consumer is an integration
test.

Deprecate the package and mark it for removal in the next release.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>